feat: use static mime types
For drive document types, spreadsheet types and presentation types.
These are expected to be valid mime types, therefore we can call
`expect` in a `LazyLock`. Unfortunately we are not able to create `Mime`
instances in const context (the `mime` crates can do that because it has
access to private fields), but at least we can avoid parsing these
values every time and, moreover, we are able to return a `Mime` instead
of an `Option<Mime>` from `DocType::mime`.