fix: issue with temporary file deleted too early
The function `file_helper::open_file` was behaving in a dangerous way:
in case the `path` was `None` a temporary file is created through
`stdin_to_file`. This file is being deleted by the `Drop` implementation
of the `Temp` structure, however this was not returned with the open
`std::fs::File` handle and the `PathBuf`, therefore once the function
returned, the handle was pointing to a file that was being deleted (at
least in theory).

Now we create an abstraction that brings the `Temp` instance in case
that branch is being taken in order to avoid the deletion of the
temporary file.

This only confirms my feelings that the abstraction provided by the
`mktemp` crate is extremely brittle, and that we should consider
replacing the functionality with something that provides a better way to
handle temporary files and directories.