feat: avoid raw copies for parent-children
In `file_tree::Folder`, the children were cloning the parent in order to
create this set of _partial_ back-references. For _partial_ I mean that
they were different objects, and this can be misleading other than being
extremely inefficient.

A new approach is to have an auxiliary `FolderInfo` that contains all
the info of the folder except for its _children_, and use an `Arc` in
order to have a real back-reference from the `parent` field. We do not
need a `Weak` because the _main_ owner of the ref-count is the `Folder`,
and the `FolderInfo` is only linked to other `FolderInfo`s, not creating
a loop.